Move domain concepts into a sub-package
