+++ src/poprox_concepts/__init__.py
@@ -1,6 +1,6 @@
-from poprox_concepts.account import Account
-from poprox_concepts.article import Article, Entity, Mention
-from poprox_concepts.click_history import ClickHistory
+from poprox_concepts.domain.account import Account
+from poprox_concepts.domain.article import Article, Entity, Mention
+from poprox_concepts.domain.click_history import ClickHistory
 
 __all__ = [
     "Account",
