+++ src/poprox_concepts/domain/article.py
