+++ src/poprox_concepts/domain/click_history.py
