+++ src/poprox_concepts/domain/internals.py
